+++ lib/frontend/marty_egui/src/windows/cpu_control.rs
@@ -32,16 +32,32 @@
 */
 use crate::*;
 use marty_core::{
+    breakpoints::StopWatchData,
     cpu_808x::CpuAddress,
     machine::{ExecutionControl, ExecutionOperation, ExecutionState},
 };
 use std::{cell::RefCell, collections::HashMap, rc::Rc};
+
+pub struct BreakpointSet<'a> {
+    pub breakpoint: &'a str,
+    pub mem_breakpoint: &'a str,
+    pub int_breakpoint: &'a str,
+    pub sw_start: &'a str,
+    pub sw_stop: &'a str,
+}
 
 pub struct CpuControl {
     exec_control: Rc<RefCell<ExecutionControl>>,
     breakpoint: String,
     mem_breakpoint: String,
     int_breakpoint: String,
+    sw_start: String,
+    sw_stop: String,
+    sw_measurements: String,
+    sw_last_cycles: String,
+    sw_total_cycles: String,
+    sw_last_duration: String,
+    sw_total_duration: String,
     step_over_target: Option<CpuAddress>,
 }
 
@@ -52,6 +68,13 @@ impl CpuControl {
             breakpoint: String::new(),
             mem_breakpoint: String::new(),
             int_breakpoint: String::new(),
+            sw_start: String::new(),
+            sw_stop: String::new(),
+            sw_measurements: String::new(),
+            sw_last_cycles: String::new(),
+            sw_total_cycles: String::new(),
+            sw_last_duration: String::new(),
+            sw_total_duration: String::new(),
             step_over_target: None,
         }
     }
@@ -224,13 +247,80 @@ impl CpuControl {
                 }
                 ui.end_row();
             });
+
+        egui::CollapsingHeader::new("StopWatch")
+            .default_open(true)
+            .show(ui, |ui| {
+                egui::Grid::new("cpu_control_stopwatch_grid")
+                    .num_columns(2)
+                    .striped(false)
+                    .min_col_width(60.0)
+                    .show(ui, |ui| {
+                        ui.label("Start address: ");
+                        if ui.text_edit_singleline(&mut self.sw_start).changed() {
+                            events.send(GuiEvent::EditBreakpoint);
+                        };
+                        ui.end_row();
+
+                        ui.label("Stop address: ");
+                        if ui.text_edit_singleline(&mut self.sw_stop).changed() {
+                            events.send(GuiEvent::EditBreakpoint);
+                        };
+                        ui.end_row();
+                    });
+
+                ui.separator();
+
+                egui::Grid::new("cpu_control_stopwatch_grid2")
+                    .num_columns(2)
+                    .striped(false)
+                    .min_col_width(60.0)
+                    .show(ui, |ui| {
+                        ui.label("Samples: ");
+                        ui.label(&self.sw_measurements);
+                        ui.end_row();
+
+                        ui.label("Last Cycles: ");
+                        ui.label(&self.sw_last_cycles);
+                        ui.end_row();
+
+                        ui.label("Total Cycles: ");
+                        ui.label(&self.sw_total_cycles);
+                        ui.end_row();
+
+                        ui.label("Last Duration: ");
+                        ui.label(&self.sw_last_duration);
+                        ui.end_row();
+
+                        ui.label("Total Duration: ");
+                        ui.label(&self.sw_total_duration);
+                        ui.end_row();
+                    });
+            });
     }
 
     pub fn set_step_over_target(&mut self, target: Option<CpuAddress>) {
         self.step_over_target = target;
     }
 
-    pub fn get_breakpoints(&mut self) -> (&str, &str, &str) {
-        (&self.breakpoint, &self.mem_breakpoint, &self.int_breakpoint)
+    pub fn get_breakpoints(&mut self) -> BreakpointSet {
+        BreakpointSet {
+            breakpoint: &self.breakpoint,
+            mem_breakpoint: &self.mem_breakpoint,
+            int_breakpoint: &self.int_breakpoint,
+            sw_start: &self.sw_start,
+            sw_stop: &self.sw_stop,
+        }
+    }
+
+    pub fn set_stopwatch_data(&mut self, data: Vec<StopWatchData>) {
+        // Eventually support multiple stopwatches; but for now just use the first one.
+        if let Some(data) = data.first() {
+            self.sw_measurements = format!("{}", data.measurements);
+            self.sw_last_cycles = format!("{}", data.last_cycles);
+            self.sw_total_cycles = format!("{}", data.total_cycles);
+            self.sw_last_duration = format!("{:.4}", data.last_duration);
+            self.sw_total_duration = format!("{:.4}", data.total_duration);
+        }
     }
 }